Use uint for bit indexes.  Use uint instead of uint64 for vector words.

+++ bitmatrix.go
@@ -4,7 +4,7 @@ package factorlib
 // linear combinations of rows which are zero.
 type bitMatrix struct {
 	// size of matrix
-	n int
+	n uint
 
 	// ids of rows (in the order in which they were added)
 	ids []interface{}
@@ -14,7 +14,7 @@ type bitMatrix struct {
 }
 
 // Return a new matrix which can handle indexes 0 <= i < n.
-func newBitMatrix(n int) *bitMatrix {
+func newBitMatrix(n uint) *bitMatrix {
 	return &bitMatrix{n, make([]interface{}, 0, n), make([]row, 0, n)}
 }
 
@@ -25,20 +25,20 @@ type row struct {
 	bits bitVec
 
 	// The column that we pivot with. == bits.firstBit()
-	pivot int
+	pivot uint
 }
 
 // Adds the vector with the given set indexes (indexes may appear multiple
 // times - an index is set if it appears an odd number of times).
 // If there is a linear combination of the added rows that xor to the zero
 // vector, addRow returns the identities of those vectors.  Otherwise returns nil.
-func (m *bitMatrix) addRow(idxs []int, id interface{}) []interface{} {
+func (m *bitMatrix) addRow(idxs []uint, id interface{}) []interface{} {
 	m.ids = append(m.ids, id)
 	bits := newBitVec(2*m.n+1)
 	for _, i := range idxs {
 		bits.toggleBit(i)
 	}
-	bits.setBit(m.n+len(m.rows))
+	bits.setBit(m.n+uint(len(m.rows)))
 	for _, r := range m.rows {
 		if bits.getBit(r.pivot) {
 			bits.xor(r.bits)
@@ -52,7 +52,7 @@ func (m *bitMatrix) addRow(idxs []int, id interface{}) []interface{} {
 
 	// found a linear combination of vectors that generates the 0 vector.
 	a := []interface{}{id}
-	for i := 0; i < m.n; i++ {
+	for i := uint(0); i < m.n; i++ {
 		if bits.getBit(m.n + i) {
 			a = append(a, m.ids[i])
 		}
